refactor(server): extract TLS / ACME helpers into server/tlscert
Step 4 of the binary-split plan. Moves the TLS-cert and ACME (HTTP-01,
DNS-01) helpers out of the server/ package root into a focused
server/tlscert subpackage with no dependency on github.com/duckdb/duckdb-go.

Files moved:
  server/certs.go      → server/tlscert/certs.go
  server/acme.go       → server/tlscert/acme.go
  server/acme_dns.go   → server/tlscert/acme_dns.go
  server/acme_test.go  → server/tlscert/acme_test.go
  server/acme_dns_test.go → server/tlscert/acme_dns_test.go

Backward compatibility preserved via type aliases and re-export `var`s in
server/tlscert_aliases.go:

  server.ACMEManager    is now an alias for tlscert.ACMEManager
  server.ACMEDNSManager is now an alias for tlscert.ACMEDNSManager
  server.NewACMEManager / NewACMEDNSManager / EnsureCertificates
    are re-export `var`s pointing at tlscert.X

So existing references in main.go, tests/integration/harness.go,
controlplane/control.go, and the Server struct's acmeManager /
acmeDNSManager fields all compile unchanged. New code should import
server/tlscert directly.

The previously-private generateSelfSignedCert is now exported as
tlscert.GenerateSelfSignedCert because server_test.go calls it
directly to generate a cert pair without going through the
EnsureCertificates file-existence check.

PR #4 originally targeted querylog + checkpoint extraction. That hit
a chicken-and-egg with DuckLakeConfig: those constructors take the
server.Config struct, and the duckdb-bound config types haven't moved
to a leaf package yet on this branch (PR #1.5 does that on a parallel
branch). Pivoted to TLS/ACME as a self-contained alternative — same
shape as PR #3, no cycle issues. The querylog / checkpoint move will
land in PR #5 once the config types are sorted.

Verified:
  - go build ./... clean
  - go build -tags kubernetes ./... clean
  - go test -short ./server/tlscert/... ./server/... ./controlplane/... ./
    all green
  - go list -deps ./server/tlscert | grep duckdb-go is empty
