+++ server/tlscert/acme_dns.go
@@ -1,4 +1,4 @@
-package server
+package tlscert
 
 import (
 	"context"
