+++ server/tlscert/acme_dns_test.go
@@ -1,4 +1,4 @@
-package server
+package tlscert
 
 import (
 	"os"
