+++ server/tlscert/acme_test.go
@@ -1,4 +1,4 @@
-package server
+package tlscert
 
 import (
 	"os"
